+++ app/RpsCompetition/Admin/Admin.php
@@ -11,6 +11,8 @@ use RpsCompetition\Constants;
 use RpsCompetition\Db\QueryCompetitions;
 use RpsCompetition\Db\QueryEntries;
 use RpsCompetition\Db\RpsDb;
+use RpsCompetition\Entity\Db\Competition;
+use RpsCompetition\Entity\Db\Entry;
 use RpsCompetition\Entries\ListTable as EntriesListTable;
 use RpsCompetition\Helpers\CommonHelper;
 use RpsCompetition\Helpers\PhotoHelper;
@@ -789,7 +791,7 @@ final class Admin
         $this->printDatepickerDefaults();
         echo '<script type="text/javascript">' . "\n";
         echo 'jQuery(function($) {' . "\n";
-        echo '	$( "#date" ).datepicker();' . "\n";
+        echo '	$( "#date" ).datepicker({dateFormat: "yy-mm-dd"});' . "\n";
         echo '});', "\n";
         echo '</script>';
         $this->displayAdminFooter();
@@ -907,10 +909,11 @@ final class Admin
         $this->displayAdminHeader('Edit Competition');
 
         if ($this->request->has('update')) {
-            echo '<div id="message" class="updated">';
             if ($updated) {
+                echo '<div id="message" class="notice notice-success">';
                 echo '<p><strong>Competition updated.</strong></p>';
             } else {
+                echo '<div id="message" class="notice notice-error">';
                 echo '<p><strong>Competition not updated.</strong></p>';
             }
             if ($wp_http_referer) {
@@ -1027,8 +1030,8 @@ final class Admin
         $this->printDatepickerDefaults();
         echo '<script type="text/javascript">' . "\n";
         echo 'jQuery(function($) {' . "\n";
-        echo '	$( "#date" ).datepicker();' . "\n";
-        echo '	$( "#close-date" ).datepicker();' . "\n";
+        echo '	$( "#date" ).datepicker({dateFormat: "yy-mm-dd"});' . "\n";
+        echo '	$( "#close-date" ).datepicker({dateFormat: "yy-mm-dd"});' . "\n";
         echo '});', "\n";
         echo '</script>';
 
@@ -1277,11 +1280,16 @@ final class Admin
         $this->displayAdminHeader('Edit Entry');
 
         if ($this->request->has('update')) {
-            echo '<div id="message" class="updated">';
-            if ($updated) {
-                echo '<p><strong>Entry updated.</strong></p>';
+
+            if (is_wp_error($updated)) {
+                echo '<div id="message" class="notice notice-error">';
+                /** @var \WP_Error $updated */
+                echo '<p><strong>Entry not updated.</strong>' .
+                     $updated->get_error_message('rpsAdminEditEntry') .
+                     '</p>';
             } else {
-                echo '<p><strong>Entry not updated.</strong></p>';
+                echo '<div id="message" class="notice notice-success">';
+                echo '<p><strong>Entry updated.</strong></p>';
             }
             if ($wp_http_referer) {
                 echo '<p><a href="' . esc_url($wp_http_referer) . '">&larr; Back to Entries</a></p>';
@@ -1319,6 +1327,17 @@ final class Admin
         echo $formBuilder->outputField($formBuilder->select('classification',
                                                             $classification_array,
                                                             $selectedClassification,
+                                                            ['autocomplete' => 'off']));
+
+        echo $formBuilder->outputLabel($formBuilder->label('score', 'Score'));
+        echo $formBuilder->outputField($formBuilder->text('score', $entry->Score));
+
+        $awards        = Constants::getAwards();
+        $selectedAward = array_search($entry->Award, $awards);
+        echo $formBuilder->outputLabel($formBuilder->label('award', 'Award'));
+        echo $formBuilder->outputField($formBuilder->select('award',
+                                                            $awards,
+                                                            $selectedAward,
                                                             ['autocomplete' => 'off']));
 
         echo $formBuilder->closeTable();
@@ -1402,14 +1421,14 @@ final class Admin
     /**
      * Perform the actual update of an entry.
      *
-     * @param array             $formOptionsNew
-     * @param int               $id
-     * @param QueryEntries      $entry       Entry record
-     * @param QueryCompetitions $competition Competition record
+     * @param array       $formOptionsNew
+     * @param int         $id
+     * @param Entry       $entry
+     * @param Competition $competition Competition record
      *
-     * @return \WP_Error|bool
+     * @return bool|\WP_Error
      */
-    private function doUpdateEntry($formOptionsNew, $id, QueryEntries $entry, QueryCompetitions $competition)
+    private function doUpdateEntry($formOptionsNew, $id, Entry $entry, Competition $competition)
     {
         $query_entries      = new QueryEntries($this->rpsdb);
         $query_competitions = new QueryCompetitions($this->rpsdb);
@@ -1427,14 +1446,17 @@ final class Admin
         $full_server_path     = $this->request->server('DOCUMENT_ROOT') . $relative_server_path;
         $dest_name            = sanitize_file_name($formOptionsNew['title']) . '+' . $user->user_login . '+' . time();
 
-        $new_competition          = $query_competitions->getCompetitionByDateClassMedium($competition->Competition_Date,
-                                                                                         $classification_array[$formOptionsNew['classification']],
-                                                                                         $medium_array[$formOptionsNew['medium']]);
+        $new_competition = $query_competitions->getCompetitionByDateClassMedium($competition->Competition_Date,
+                                                                                $classification_array[$formOptionsNew['classification']],
+                                                                                $medium_array[$formOptionsNew['medium']]);
+
         $data                     = [];
         $data['Competition_ID']   = $new_competition->ID;
         $data['ID']               = $id;
         $data['Server_File_Name'] = $relative_server_path . '/' . $dest_name . '.jpg';
         $data['Title']            = $formOptionsNew['title'];
+        $data['Score']            = $formOptionsNew['score'];
+        $data['Award']            = $formOptionsNew['award'];
 
         // Need to create the destination folder?
         CommonHelper::createDirectory($full_server_path);
@@ -1667,6 +1689,79 @@ final class Admin
     }
 
     /**
+     * Perform check to see if the updated entry is valid.
+     *
+     * @param array       $formOptions
+     * @param array       $formOptionsNew
+     * @param Entry       $entry       Entry record
+     * @param Competition $competition Competition record
+     *
+     * @return bool|\WP_Error
+     */
+    private function isValidEntry($formOptions, $formOptionsNew, Entry $entry, $competition)
+    {
+        $query_entries        = new QueryEntries($this->rpsdb);
+        $query_competitions   = new QueryCompetitions($this->rpsdb);
+        $medium_array         = Constants::getMediums();
+        $classification_array = Constants::getClassifications();
+        $user                 = get_user_by('id', $entry->Member_ID);
+
+        if ($formOptions['medium'] !== $formOptionsNew['medium'] ||
+            $formOptions['classification'] !== $formOptionsNew['classification']
+        ) {
+            $new_competition = $query_competitions->getCompetitionByDateClassMedium($competition->Competition_Date,
+                                                                                    $classification_array[$formOptionsNew['classification']],
+                                                                                    $medium_array[$formOptionsNew['medium']]);
+
+            $max_per_id = $query_entries->countEntriesByCompetitionId($new_competition->ID, $user->ID);
+            if ($max_per_id >= $new_competition->Max_Entries) {
+                $error_message = 'The maximum of ' .
+                                 $new_competition->Max_Entries .
+                                 ' entries into this competition has been reached. Update cancelled.';
+
+                return new \WP_Error('rpsAdminEditEntry', $error_message);
+            }
+        }
+
+        $awards_array = Constants::getAwards();
+        if (!array_key_exists($formOptionsNew['award'], $awards_array)) {
+            $error_message = 'Invalid Award. Update cancelled.';
+
+            return new \WP_Error('rpsAdminEditEntry', $error_message);
+        }
+        $return = $this->isValidScore($formOptionsNew);
+
+        unset($query_entries, $query_competitions);
+
+        return $return;
+    }
+
+    /**
+     * @param $formOptionsNew
+     *
+     * @return bool|\WP_Error
+     */
+    private function isValidScore($formOptionsNew)
+    {
+        if (is_null($formOptionsNew['score'])) {
+            return true;
+        }
+        if (!is_numeric($formOptionsNew['score'])) {
+            $error_message = 'The score has to be numeric. Update cancelled.';
+
+            return new \WP_Error('rpsAdminEditEntry', $error_message);
+        }
+
+        if ($formOptionsNew['score'] < 6 || $formOptionsNew['score'] > 9) {
+            $error_message = 'The score has to be between 6 and 9. Update cancelled.';
+
+            return new \WP_Error('rpsAdminEditEntry', $error_message);
+        }
+
+        return true;
+    }
+
+    /**
      * Display plugin Copyright
      */
     private function printAdminFooter()
@@ -1687,7 +1782,6 @@ final class Admin
         echo '<script type="text/javascript">' . "\n";
         echo 'jQuery(function($) {' . "\n";
         echo ' $.datepicker.setDefaults({' . "\n";
-        echo '   dateFormat: \'yy-mm-dd\', ' . "\n";
         echo '   showButtonPanel: true, ' . "\n";
         echo '   buttonImageOnly: true, ' . "\n";
         echo '   buttonImage: "' . CommonHelper::getPluginUrl('calendar.png',
@@ -1757,7 +1851,6 @@ final class Admin
         $formOptions = $this->request->input('entry-edit');
         $id          = (int) $this->request->input('entry');
         $entry       = $query_entries->getEntryById($id);
-        /** @var QueryCompetitions $competition */
         $competition = $competition_query->getCompetitionById($entry->Competition_ID);
 
         $medium_array         = Constants::getMediums();
@@ -1770,8 +1863,13 @@ final class Admin
         $formOptionsNew['title']          = empty($formOptions['title']) ? $entry->Title : $formOptions['title'];
         $formOptionsNew['medium']         = empty($formOptions['medium']) ? $selectedMedium : $formOptions['medium'];
         $formOptionsNew['classification'] = empty($formOptions['classification']) ? $selectedClassification : $formOptions['classification'];
+        $formOptionsNew['score']          = empty($formOptions['score']) ? null : $formOptions['score'];
+        $formOptionsNew['award']          = $formOptions['award'];
 
-        $return = $this->doUpdateEntry($formOptionsNew, $id, $entry, $competition);
+        $return = $this->isValidEntry($formOptions, $formOptionsNew, $entry, $competition);
+        if (!is_wp_error($return)) {
+            $return = $this->doUpdateEntry($formOptionsNew, $id, $entry, $competition);
+        }
 
         unset($query_entries);