+++ app/RpsCompetition/Constants.php
@@ -37,8 +37,19 @@ final class Constants
      */
     const PLUGIN_FILE           = 'avh-rps-competition/avh-rps-competition.php';
     const PLUGIN_README_URL     = '';
-    const PLUGIN_VERSION        = '3.0.1';
+    const PLUGIN_VERSION        = '3.0.2';
     const SLUG_COMPETITION_EDIT = 'avh-rps-competition-edit';
+
+    public static function getAwards()
+    {
+        return [
+            '1st' => '1st',
+            '2nd' => '2nd',
+            '3rd' => '3rd',
+            'HM'  => 'HM',
+            null  => 'None'
+        ];
+    }
 
     /**
      * Return an array with the available classifications.