+++ app/RpsCompetition/Frontend/Requests/BanquetEntries/RequestBanquetEntries.php
@@ -78,11 +78,10 @@ class RequestBanquetEntries
                      ->isClicked()
             ) {
                 $this->handleUpdate();
+                wp_redirect($this->request->fullUrl());
+                exit();
             }
         }
-
-        wp_redirect($this->request->fullUrl());
-        exit();
     }
 
     /**